pyproject.toml
[tool.black]
target-version = ['py310']
line-length = 100
include = '\.pyi?$'

[tool.isort]
profile="black"
line_length=100

[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = "fastapi test project"
authors = ["JngMkk"]
readme = "README.md"
packages = [{include = "fastapi_project"}]

[tool.poetry.dependencies]
python = "^3.10"
pymysql = "^1.0.3"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
pydantic = {extras = ["email"], version = "^1.10.7"}
cryptography = "^40.0.1"
sqlmodel = "^0.0.8"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
